Moving raw plugin import, updating Gradle file.
